Demographics Simulation
This is a little experiment and project that I made to learn JavaScript. It was partially aided by AI, mostly fixing syntax and it giving me suggestions. It helped me learn quite a lot. The code is an absolute mess and I came to a point where I barely know what the actual fuck is going on. But it functions. It needs to be tweaked, optimized and improved in all aspects. And it could probably be simplified. But to hell with it, I am uploading this shit here.